requirements.txt
playwright==1.58.0
numpy==2.3.0
duckdb==1.5.1
pandas==3.0.1
pypdf==6.9.2
pytesseract==0.3.13
Pillow>=11.0.0
apache-airflow==3.0.6

# this is for superset
duckdb-engine==0.17.0